backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-multipart==0.0.12
openpyxl==3.1.5
PyJWT==2.9.0
pydantic-settings==2.6.1
boto3==1.35.24
httpx==0.27.2
